fix: poolBalancesMixin
